Refactor

+++ hasher.py
@@ -2,6 +2,7 @@ import os
 import sys
 import glob
 import hashlib
+import argparse
 
 
 TXT_OR_LINK = 'link'
@@ -15,7 +16,7 @@ def md5(fname):
 
 
 class Consolidator(object):
-    def __init__(self, old_path, new_path):
+    def __init__(self, old_path, new_path, dry_run=False):
         self.new_path = new_path
         self.old_path = old_path
         if self.old_path.endswith('/') and not self.new_path.endswith('/'):
@@ -23,50 +24,72 @@ class Consolidator(object):
         if not self.old_path.endswith('/') and self.new_path.endswith('/'):
             self.old_path += '/'
 
+        self.dry_run = dry_run
+
     def do_it(self):
+        print("Comparing:")
+        print("Old path: ", self.old_path)
+        print("New path: ", self.new_path)
         files = glob.glob(self.new_path + '/**/*', recursive=True)
         for f in files:
             if os.path.isfile(f):
                 self.check_file(f)
 
     def check_file(self, file_path):
-        new_hash = md5(file_path)
+        new_source_file, new_hash = self.get_source_file_and_hash(file_path)
         old_file_path = file_path.replace(self.new_path, self.old_path)
-        old_hash = None
-        prev_file = None
-        # either txt or symlink
-        # if old file is hash txt file...
-        already_hashed_path = old_file_path + '.txt'
-        if os.path.exists(already_hashed_path):
-            with open(already_hashed_path, 'r') as old_hashed_file:
-                content = old_hashed_file.read().strip()
-                old_hash, prev_file = content.split('||')
-        elif os.path.islink(old_file_path):
-            prev_file = os.path.realpath(old_file_path)
-            old_hash = md5(prev_file)
-        if old_hash is None and os.path.exists(old_file_path):
-            old_hash = md5(old_file_path)
+        old_source_file, old_hash = self.get_source_file_and_hash(file_path)
 
-        if old_hash is not None and new_hash == old_hash:
-            if prev_file is None:
-                prev_file = old_file_path
-            print('deleting:', file_path)
-            os.remove(file_path)
-            # either txt or link option
-            if TXT_OR_LINK == 'txt':
-                with open(file_path + '.txt', 'w') as new_dupe_file:
-                    new_dupe_file.write(new_hash + '||' + prev_file + '\n')
+        # if there is an old file, hashes are the same, and source files aren't already equal:
+        if old_hash is not None and new_hash == old_hash and new_source_file != old_source_file:
+            if not self.dry_run:
+                print('Deleting:', file_path)
+                os.remove(file_path)
+                # either txt or link option
+                if TXT_OR_LINK == 'txt':
+                    with open(file_path + '.txt', 'w') as new_dupe_file:
+                        new_dupe_file.write(new_hash + '||' + old_source_file + '\n')
+                else:
+                    os.symlink(old_source_file, file_path)
             else:
-                os.symlink(prev_file, file_path)
+                print("Would have deleted: ", file_path)
+
+    @staticmethod
+    def get_source_file_and_hash(path):
+        original_source = None
+        hash = None
+        possible_txt_path = path + '.txt'
+        if os.path.islink(path):
+            original_source = os.path.realpath(path)
+            hash = md5(original_source)
+        elif os.path.exists(possible_txt_path):
+            with open(possible_txt_path, 'r') as hashed_file:
+                content = hashed_file.read().strip()
+                hash, original_source = content.split('||')
+        elif os.path.exists(path):
+            hash = md5(path)
+            original_source = path
+        return original_source, hash
+
 
 if __name__ == '__main__':
-    old_path = sys.argv[1]
+    description = "Tool to consolidate OP-1 backups"
+    parser = argparse.ArgumentParser(description=description)
+    parser.add_argument("-p", "--previous", help="directory of previous backup", required=True)
+    parser.add_argument("-n", "--new", help="directory of newer backup", required=True)
+    parser.add_argument('--modify-original', dest='in_place', action='store_true')
+    parser.set_defaults(in_place=False)
+    parser.add_argument('--dry-run', dest='dry_run', action='store_true')
+    parser.set_defaults(dry_run=False)
+    parsed = parser.parse_args()
+
+    old_path = parsed.previous
     if not old_path.startswith('/'):
         old_path = os.path.realpath(os.path.join(os.getcwd(), old_path))
-    new_path = sys.argv[2]
+    new_path = parsed.new
     if not new_path.startswith('/'):
         new_path = os.path.realpath(os.path.join(os.getcwd(), new_path))
 
-    c = Consolidator(old_path, new_path)
+    c = Consolidator(old_path, new_path, dry_run=parsed.dry_run)
 
     c.do_it()
